hira can use modules that arent defined in same crate
this is pretty cool. basically module writers can write their code exactly the same way regardless of if its going to be used in the final executable or not. and users who depend on modules can simply load them via cargo the normal way and it just works TM. the way it works is: cargo compiles dependencies first. and when cargo compiled the dependent modules, it still uses hira's proc macro, and since we detect they are level2 modules, we cache them to disk. then finally when cargo compiles the final binary, the user's code, it doesnt see the lvl2 modules in memory, so it tries to load them from disk which succeeds because they were previously parsed when cargo was compiling the dependencies